refactor: refactor naming of database batch exports
